requirements/development.txt
-i https://pypi.python.org/simple
-r base.txt

django-debug-toolbar==3.4.0
sqlparse==0.4.2
werkzeug==2.1.2
